feat(MessageManager): update to TypeScript
